+++ fe/fe-core/src/main/java/org/apache/doris/common/proc/ClusterHealthProcDir.java
@@ -17,30 +17,13 @@
 
 package org.apache.doris.common.proc;
 
-import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
 import org.apache.doris.catalog.Catalog;
-import org.apache.doris.catalog.ColocateTableIndex;
-import org.apache.doris.catalog.Database;
-import org.apache.doris.catalog.MaterializedIndex;
-import org.apache.doris.catalog.OlapTable;
-import org.apache.doris.catalog.Partition;
-import org.apache.doris.catalog.ReplicaAllocation;
-import org.apache.doris.catalog.Table;
-import org.apache.doris.catalog.Tablet;
-import org.apache.doris.clone.TabletSchedCtx;
 import org.apache.doris.common.AnalysisException;
-import org.apache.doris.common.Pair;
-import org.apache.doris.system.SystemInfoService;
 
 import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Objects;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /*
  * show proc "/cluster_health";
@@ -48,18 +31,9 @@ import java.util.stream.Stream;
 
 public class ClusterHealthProcDir implements ProcDirInterface {
     public static final ImmutableList<String> TITLE_NAMES = new ImmutableList.Builder<String>()
-            .add("DbId").add("DbName").add("TabletNum").add("HealthyNum").add("ReplicaMissingNum")
-            .add("VersionIncompleteNum").add("ReplicaRelocatingNum").add("RedundantNum")
-            .add("ReplicaMissingInClusterNum").add("ReplicaMissingForTagNum")
-            .add("ForceRedundantNum").add("ColocateMismatchNum").add("ColocateRedundantNum")
-            .add("NeedFurtherRepairNum").add("UnrecoverableNum").add("ReplicaCompactionTooSlowNum")
-            .build();
+            .add("Items").build();
 
-    private Catalog catalog;
-
-    public ClusterHealthProcDir(Catalog catalog) {
-        Preconditions.checkNotNull(catalog);
-        this.catalog = catalog;
+    public ClusterHealthProcDir() {
     }
 
     @Override
@@ -68,226 +42,17 @@ public class ClusterHealthProcDir implements ProcDirInterface {
     }
 
     @Override
-    public ProcNodeInterface lookup(String dbIdStr) throws AnalysisException {
-        try {
-            long dbId = Long.parseLong(dbIdStr);
-            return catalog.getDb(dbId).map(IncompleteUnhealthyTabletsProcNode::new).orElse(null);
-        } catch (NumberFormatException e) {
-            throw new AnalysisException("Invalid db id format: " + dbIdStr);
+    public ProcNodeInterface lookup(String itemStr) {
+        if (itemStr.equalsIgnoreCase("tablet_health")) {
+            return new TabletHealthProcDir(Catalog.getCurrentCatalog());
         }
+        return null;
     }
 
     @Override
     public ProcResult fetchResult() throws AnalysisException {
-        List<DBTabletStatistic> statistics = catalog.getDbIds().parallelStream()
-                // skip information_schema database
-                .flatMap(id -> Stream.of(id == 0 ? null : catalog.getDbNullable(id)))
-                .filter(Objects::nonNull).map(DBTabletStatistic::new)
-                // sort by dbName
-                .sorted(Comparator.comparing(db -> db.db.getFullName()))
-                .collect(Collectors.toList());
-
-        List<List<String>> rows = new ArrayList<>(statistics.size() + 1);
-        for (DBTabletStatistic statistic : statistics) {
-            rows.add(statistic.toRow());
-        }
-        rows.add(statistics.stream().reduce(new DBTabletStatistic(), DBTabletStatistic::reduce).toRow());
+        List<List<String>> rows = new ArrayList<>(1);
+        rows.add(Lists.newArrayList("tablet_health"));
         return new BaseProcResult(TITLE_NAMES, rows);
-    }
-
-    static class DBTabletStatistic {
-        boolean summary;
-        Database db;
-        int dbNum;
-        int tabletNum;
-        int healthyNum;
-        int replicaMissingNum;
-        int versionIncompleteNum;
-        int replicaRelocatingNum;
-        int redundantNum;
-        int replicaMissingInClusterNum;
-        int replicaMissingForTagNum;
-        int forceRedundantNum;
-        int colocateMismatchNum;
-        int colocateRedundantNum;
-        int needFurtherRepairNum;
-        int unrecoverableNum;
-        int replicaCompactionTooSlowNum;
-
-        Set<Long> replicaMissingTabletIds;
-        Set<Long> versionIncompleteTabletIds;
-        Set<Long> replicaRelocatingTabletIds;
-        Set<Long> redundantTabletIds;
-        Set<Long> replicaMissingInClusterTabletIds;
-        Set<Long> replicaMissingForTagTabletIds;
-        Set<Long> forceRedundantTabletIds;
-        Set<Long> colocateMismatchTabletIds;
-        Set<Long> colocateRedundantTabletIds;
-        Set<Long> needFurtherRepairTabletIds;
-        Set<Long> unrecoverableTabletIds;
-        Set<Long> replicaCompactionTooSlowTabletIds;
-
-        DBTabletStatistic() {
-            this.summary = true;
-        }
-
-        DBTabletStatistic(Database db) {
-            Preconditions.checkNotNull(db);
-            this.summary = false;
-            this.db = db;
-            this.dbNum = 1;
-            this.replicaMissingTabletIds = new HashSet<>();
-            this.versionIncompleteTabletIds = new HashSet<>();
-            this.replicaRelocatingTabletIds = new HashSet<>();
-            this.redundantTabletIds = new HashSet<>();
-            this.replicaMissingInClusterTabletIds = new HashSet<>();
-            this.replicaMissingForTagTabletIds = new HashSet<>();
-            this.forceRedundantTabletIds = new HashSet<>();
-            this.colocateMismatchTabletIds = new HashSet<>();
-            this.colocateRedundantTabletIds = new HashSet<>();
-            this.needFurtherRepairTabletIds = new HashSet<>();
-            this.unrecoverableTabletIds = new HashSet<>();
-            this.replicaCompactionTooSlowTabletIds = new HashSet<>();
-
-            SystemInfoService infoService = Catalog.getCurrentSystemInfo();
-            ColocateTableIndex colocateTableIndex = Catalog.getCurrentColocateIndex();
-            List<Long> aliveBeIdsInCluster = infoService.getClusterBackendIds(db.getClusterName(), true);
-            db.getTables().stream().filter(t -> t != null && t.getType() == Table.TableType.OLAP).forEach(t -> {
-                OlapTable olapTable = (OlapTable) t;
-                ColocateTableIndex.GroupId groupId = colocateTableIndex.isColocateTable(olapTable.getId()) ?
-                        colocateTableIndex.getGroup(olapTable.getId()) : null;
-                olapTable.readLock();
-                try {
-                    for (Partition partition : olapTable.getAllPartitions()) {
-                        ReplicaAllocation replicaAlloc = olapTable.getPartitionInfo().getReplicaAllocation(partition.getId());
-                        for (MaterializedIndex materializedIndex : partition.getMaterializedIndices(MaterializedIndex.IndexExtState.VISIBLE)) {
-                            List<Tablet> tablets = materializedIndex.getTablets();
-                            for (int i = 0; i < tablets.size(); ++i) {
-                                Tablet tablet = tablets.get(i);
-                                ++tabletNum;
-                                Tablet.TabletStatus res = null;
-                                if (groupId != null) {
-                                    Set<Long> backendsSet = colocateTableIndex.getTabletBackendsByGroup(groupId, i);
-                                    res = tablet.getColocateHealthStatus(partition.getVisibleVersion(), replicaAlloc, backendsSet);
-                                } else {
-                                    Pair<Tablet.TabletStatus, TabletSchedCtx.Priority> pair = tablet.getHealthStatusWithPriority(
-                                            infoService, db.getClusterName(),
-                                            partition.getVisibleVersion(),
-                                            replicaAlloc, aliveBeIdsInCluster);
-                                    res = pair.first;
-                                }
-                                switch (res) {
-                                    case HEALTHY:
-                                        healthyNum++;
-                                        break;
-                                    case REPLICA_MISSING:
-                                        replicaMissingNum++;
-                                        replicaMissingTabletIds.add(tablet.getId());
-                                        break;
-                                    case VERSION_INCOMPLETE:
-                                        versionIncompleteNum++;
-                                        versionIncompleteTabletIds.add(tablet.getId());
-                                        break;
-                                    case REPLICA_RELOCATING:
-                                        replicaRelocatingNum++;
-                                        replicaRelocatingTabletIds.add(tablet.getId());
-                                        break;
-                                    case REDUNDANT:
-                                        redundantNum++;
-                                        redundantTabletIds.add(tablet.getId());
-                                        break;
-                                    case REPLICA_MISSING_IN_CLUSTER:
-                                        replicaMissingInClusterNum++;
-                                        replicaMissingInClusterTabletIds.add(tablet.getId());
-                                        break;
-                                    case REPLICA_MISSING_FOR_TAG:
-                                        replicaMissingForTagNum++;
-                                        replicaMissingForTagTabletIds.add(tablet.getId());
-                                        break;
-                                    case FORCE_REDUNDANT:
-                                        forceRedundantNum++;
-                                        forceRedundantTabletIds.add(tablet.getId());
-                                        break;
-                                    case COLOCATE_MISMATCH:
-                                        colocateMismatchNum++;
-                                        colocateMismatchTabletIds.add(tablet.getId());
-                                        break;
-                                    case COLOCATE_REDUNDANT:
-                                        colocateRedundantNum++;
-                                        colocateRedundantTabletIds.add(tablet.getId());
-                                        break;
-                                    case NEED_FURTHER_REPAIR:
-                                        needFurtherRepairNum++;
-                                        needFurtherRepairTabletIds.add(tablet.getId());
-                                        break;
-                                    case UNRECOVERABLE:
-                                        unrecoverableNum++;
-                                        unrecoverableTabletIds.add(tablet.getId());
-                                        break;
-                                    case REPLICA_COMPACTION_TOO_SLOW:
-                                        replicaCompactionTooSlowNum++;
-                                        replicaCompactionTooSlowTabletIds.add(tablet.getId());
-                                        break;
-                                }
-                            } // end for tablets
-                        } // end for indices
-                    } // end for partitions
-                } finally {
-                    olapTable.readUnlock();
-                }
-            });
-        }
-
-        DBTabletStatistic reduce(DBTabletStatistic other) {
-            if (this.summary) {
-                this.dbNum += other.dbNum;
-                this.tabletNum += other.tabletNum;
-                this.healthyNum += other.healthyNum;
-                this.replicaMissingNum += other.replicaMissingNum;
-                this.versionIncompleteNum += other.versionIncompleteNum;
-                this.replicaRelocatingNum += other.replicaRelocatingNum;
-                this.redundantNum += other.redundantNum;
-                this.replicaMissingInClusterNum += other.replicaMissingInClusterNum;
-                this.replicaMissingForTagNum += other.replicaMissingForTagNum;
-                this.forceRedundantNum += other.forceRedundantNum;
-                this.colocateMismatchNum += other.colocateMismatchNum;
-                this.colocateRedundantNum += other.colocateRedundantNum;
-                this.needFurtherRepairNum += other.needFurtherRepairNum;
-                this.unrecoverableNum += unrecoverableNum;
-                this.replicaCompactionTooSlowNum += replicaCompactionTooSlowNum;
-                return this;
-            } else if (other.summary) {
-                return other.reduce(this);
-            } else {
-                return new DBTabletStatistic().reduce(this).reduce(other);
-            }
-        }
-
-        List<String> toRow() {
-            List<Object> row = new ArrayList<>(TITLE_NAMES.size());
-            if (summary) {
-                row.add("Total");
-                row.add(dbNum);
-            } else {
-                row.add(db.getId());
-                row.add(db.getFullName());
-            }
-
-            row.add(tabletNum);
-            row.add(healthyNum);
-            row.add(replicaMissingNum);
-            row.add(versionIncompleteNum);
-            row.add(replicaRelocatingNum);
-            row.add(redundantNum);
-            row.add(replicaMissingInClusterNum);
-            row.add(replicaMissingForTagNum);
-            row.add(forceRedundantNum);
-            row.add(colocateMismatchNum);
-            row.add(colocateRedundantNum);
-            row.add(needFurtherRepairNum);
-            row.add(unrecoverableNum);
-            row.add(replicaCompactionTooSlowNum);
-            return row.stream().map(String::valueOf).collect(Collectors.toList());
-        }
     }
 }
