+++ fe/fe-core/src/main/java/org/apache/doris/common/proc/IncompleteTabletsProcNode.java
@@ -28,8 +28,11 @@ import java.util.Collections;
 
 public class IncompleteTabletsProcNode implements ProcNodeInterface {
     public static final ImmutableList<String> TITLE_NAMES = new ImmutableList.Builder<String>()
-            .add("UnhealthyTablets").add("InconsistentTablets").add("CloningTablets").add("BadTablets")
-            .add("CompactionTooSlowTablets").add("OversizeTablets")
+            .add("ReplicaMissingTablets").add("VersionIncompleteTablets").add("ReplicaRelocatingTablets")
+            .add("RedundantTablets").add("ReplicaMissingInClusterTablets").add("ReplicaMissingForTagTablets")
+            .add("ForceRedundantTablets").add("ColocateMismatchTablets").add("ColocateRedundantTablets")
+            .add("NeedFurtherRepairTablets").add("UnrecoverableTablets").add("ReplicaCompactionTooSlowTablets")
+            .add("InconsistentTablets").add("OversizeTablets")
             .build();
     private static final Joiner JOINER = Joiner.on(",");
 
@@ -41,13 +44,21 @@ public class IncompleteTabletsProcNode implements ProcNodeInterface {
 
     @Override
     public ProcResult fetchResult() throws AnalysisException {
-        StatisticProcDir.DBStatistic statistic = new StatisticProcDir.DBStatistic(db);
+        TabletHealthProcDir.DBTabletStatistic statistic = new TabletHealthProcDir.DBTabletStatistic(db);
         return new BaseProcResult(TITLE_NAMES, Collections.singletonList(Arrays.asList(
-                JOINER.join(statistic.unhealthyTabletIds),
-                JOINER.join(statistic.inconsistentTabletIds),
-                JOINER.join(statistic.cloningTabletIds),
+                JOINER.join(statistic.replicaMissingTabletIds),
+                JOINER.join(statistic.versionIncompleteTabletIds),
+                JOINER.join(statistic.replicaRelocatingTabletIds),
+                JOINER.join(statistic.redundantTabletIds),
+                JOINER.join(statistic.replicaMissingInClusterTabletIds),
+                JOINER.join(statistic.replicaMissingForTagTabletIds),
+                JOINER.join(statistic.forceRedundantTabletIds),
+                JOINER.join(statistic.colocateMismatchTabletIds),
+                JOINER.join(statistic.colocateRedundantTabletIds),
+                JOINER.join(statistic.needFurtherRepairTabletIds),
                 JOINER.join(statistic.unrecoverableTabletIds),
-                JOINER.join(statistic.compactionTooSlowTabletIds),
+                JOINER.join(statistic.replicaCompactionTooSlowTabletIds),
+                JOINER.join(statistic.inconsistentTabletIds),
                 JOINER.join(statistic.oversizeTabletIds)
         )));
     }
